requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1
django-braces==1.15.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
dnspython==2.3.0
mysqlclient==2.1.1
Pillow==10.0.1
pymongo==4.3.3
pypdf==3.8.0
pytz==2023.3
sqlparse==0.4.3
typing-extensions==4.5.0
Unidecode==1.3.6
uWSGI==2.0.21
drf_yasg==1.21.7
requests
ldap3==2.9.1
pre-commit
flake8
black
